Refactor JSON API endpoints into a blueprint.
